version number in stamp
